`belongs_to` method now uses singular name
The `belongs_to` method was singularizing the table-names. So, if I have
a table Users and a table Cars and each user has one car, the auto-generated
code would just use the table name to name the relationship which would look
like:

  @user = User.find(1)
  @car = @user.cars       #=> should't be plural

This was fixed by calling the `singularize` method on the table name when
generating the code.