Legacy databases copying support
Columns are not added after table creation because if we are
creating table without id column create statement will fail
because list of columns is empty (seen on MySQL). Therefore we are
adding columns in the block with same logic as before.

In addition 'type' columns, if they exist, are handled correctly
avoiding setting values to nil for such columns (by using option
withouth_protection when new record is created).